Move .schedule to Commander via a Scheduling mixin
So that custom commander also get scheduling, including (later) Sourced deciders that will be able to include the mixin
