+++ lib/sidereal/app.rb
@@ -137,28 +137,9 @@ module Sidereal
         self
       end
 
-      # Register a cron-scheduled block. The block runs on every tick of
-      # the cron expression, on a fiber spawned by {Sidereal.scheduler}.
-      # Inside the block, +dispatch(MessageClass, payload)+ enqueues a
-      # command onto {Sidereal.store} stamped with
-      # +metadata: { producer: '<cron expr>' }+.
-      #
-      # @example
-      #   schedule '5 0 * * *' do
-      #     dispatch Cleanup, foo: 'bar'
-      #   end
-      #
-      # @param cron_expr [String] cron expression (5 or 6 fields)
-      # @return [self]
-      def schedule(cron_expr, &block)
-        Sidereal.scheduler.schedule(cron_expr, &block)
-
-        command Sidereal::System::TriggerSchedule do |cmd|
-          schedule = Sidereal.scheduler.find(cmd.payload.schedule_id)
-          return unless schedule
-
-          schedule.run_in(self, cmd)
-        end
+      def schedule(...)
+        commander.schedule(...)
+        Sidereal.register(commander)
         self
       end
